# perf: improve hashblob performance up to 100%