--- a/src/api/hashBlob.js
+++ b/src/api/hashBlob.js
@@ -38,14 +38,14 @@ export async function hashBlob({ object }) {
 
     // Convert object to buffer
     if (typeof object === 'string') {
-      object = Buffer.from(object, 'utf8')
-    } else {
       object = Buffer.from(object)
+    } else if (!(object instanceof Uint8Array)) {
+      object = new Uint8Array(object)
     }
 
     const type = 'blob'
     const { oid, object: _object } = await hashObject({
-      type: 'blob',
+      type,
       format: 'content',
       object,
     })