<!-- Oh wow! Thanks for opening a pull request! 😁 🎉 -->
<!-- You are very welcome here and any contribution is appreciated. 👍 -->
<!-- Choose one of the checklists if it applies to you and delete the rest. -->

## performance improvement for hashBlob

## 50k Promises (Promise.all)

|            | **before (ms)** | **after (ms)** |
| ---------- | --------------- | -------------- |
| **Buffer** | 3822.918        | 1872.766       |
| **string** | 4072.27         | 2300           |


Buffer = 104.13% faster
string = 77.06% faster

### 50k Promises sequential

|            | **before (ms)** | **after (ms)** |
| ---------- | --------------- | -------------- |
| **Buffer** | 2821.764        | 1811.034       |
| **string** | 2509.682        | 1790           |


Buffer = 55.81 % faster
string = 40.21 % faster


- hashObject return a `Uint8Array`, no need to create another.
- hashBlob skip wrapping object in `Buffer` if it is already a `Uint8Array`
- `GitObject.wrap` : Create only 1 buffer instead of 3