@lukecotter 2025-08-06
@jcubic 

I was thinking about this and I think the change in hashBlob should probably have been this to be safer.

```

if (typeof object === 'string') {
      object = Buffer.from(object, 'utf8')
    } else if (!(object instanceof Uint8Array)) {
      object = Buffer.from(object)
    }
```

That is close to the original and will support more types but still skip wrapping for Buffer and Uint8Array. Given that the only types that should be provided are string and uint8array or subclasses it's probably not a big deal.